+++ tests/resources/flux/link.blade.php
@@ -1,0 +1,2 @@
+@props(['href' => null, 'external' => false])
+<a href="{{ $href }}" data-flux="link" {{ $attributes }}>{{ $slot }}</a>